auth.pike: Remove invocation of Yosemite.py and replace with exit trap
Yosemite.py is no longer supported and will not be present. In the Yosemite
Project, this script would hold the shell until terminated. This causes the
magic state of SSHFS mentioned in the comment block, which needed to be
recreated another way. Plus, it seems easier to have a thing to close when
done with the share, than to have the user remember an incantation like
`fusermount -u /home/[user]/Videos/Merlin`.